+++ lua/configs/ui/match-visual.lua
@@ -1,0 +1,5 @@
+return {
+    min_length = 1,
+    hl_group = "Visual",
+    match_id = 118,
+}